MAINT: Extract useful info from tk gui
